pyproject.toml
[project]
name = "contabilidad"
version = "0.1.0"
description = "Sistema de contabilidad con Django"
authors = [{ name = "Daniel Mendieta", email = "danielmc0911@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.2.3",
    "psycopg2-binary==2.9.9",
    "django-htmx==1.18.0",
    "sqlparse==0.5.3",
    "asgiref==3.8.1",
    "typing-extensions==4.14.0",
    "packaging==25.0",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "black==25.1.0",
    "flake8==7.3.0",
    "isort==6.0.1",
    "mypy==1.16.1",
    "mypy-extensions==1.1.0",
    "pycodestyle==2.14.0",
    "pyflakes==3.4.0",
    "click==8.2.1",
    "platformdirs==4.3.8",
    "pathspec==0.12.1",
    "pytest==8.4.1",
    "pytest-django==4.11.1",
    "factory-boy==3.3.3"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["contabilidad*", "asientos_contables*"]
namespaces = false

[tool.uv]
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "contabilidad.settings"
python_files = "tests.py test_*.py *_tests.py"

[tool.isort]
profile = "black"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
